refactor(report): split per-renderer submodules
Mechanical split of the 2958-line src/report.rs into per-renderer
files. No behavior changes — every test (140 unit + 84 CLI + others)
still passes; render output is byte-identical.

  src/report.rs              141  dispatcher (Format, render, render_delta) + re-exports
  src/report/types.rs         81  Grade, coverage_bar, delta_display
  src/report/links.rs         88  SourceLinks, link_path, linkify
  src/report/json.rs          84  render_json, render_delta_json, Envelope, schema URLs
  src/report/github.rs       108  ::warning annotations
  src/report/human.rs        236  comfy-table renderers
  src/report/markdown.rs     223  exhaustive GFM
  src/report/pr_comment.rs   485  opinionated PR comment + LCP helpers + bucketing
  src/report/summary.rs       93  --summary aggregate output
  src/report/per_crate.rs    107  workspace rollup
  src/report/tests.rs       1468  all existing tests, kept colocated

Cross-module helpers are pub(crate); the public API (Format, render,
render_delta, render_summary, render_delta_summary, SourceLinks,
Envelope, SCHEMA_VERSION, REPORT_SCHEMA_URL, DELTA_SCHEMA_URL,
crappy_count) is preserved via re-exports — lib.rs and main.rs
unchanged.
