+++ src/report/markdown.rs
@@ -1,0 +1,223 @@
+//! `--format markdown` — exhaustive GFM table for archived artifacts.
+//!
+//! No row caps, no `<details>` collapsing. The `pr-comment` format trades
+//! completeness for readability; this format is the dual.
+
+use super::links::{SourceLinks, linkify};
+use super::per_crate::write_per_crate_markdown;
+use super::types::{Grade, delta_display};
+use super::write_pr_comment_marker;
+use crate::delta::{DeltaEntry, DeltaReport, DeltaStatus};
+use crate::merge::CrapEntry;
+use anyhow::Result;
+use std::io::Write;
+
+fn write_markdown_absolute_heading(
+    crappy: usize,
+    threshold: f64,
+    out: &mut dyn Write,
+) -> Result<()> {
+    if crappy == 0 {
+        writeln!(out, "## ✅ No CRAP threshold violations")?;
+    } else {
+        writeln!(
+            out,
+            "## ⚠️ {crappy} function(s) exceed CRAP threshold {threshold}"
+        )?;
+    }
+    writeln!(out)?;
+    Ok(())
+}
+
+fn write_markdown_absolute_summary(
+    crappy: usize,
+    total: usize,
+    threshold: f64,
+    out: &mut dyn Write,
+) -> Result<()> {
+    writeln!(out)?;
+    if crappy == 0 {
+        writeln!(
+            out,
+            "✓ {total} function(s) analyzed; none exceed CRAP threshold {threshold}."
+        )?;
+    } else {
+        writeln!(
+            out,
+            "✗ {crappy}/{total} function(s) exceed CRAP threshold {threshold}."
+        )?;
+    }
+    Ok(())
+}
+
+fn write_markdown_entries_table(
+    entries: &[CrapEntry],
+    threshold: f64,
+    links: Option<&SourceLinks>,
+    out: &mut dyn Write,
+) -> Result<()> {
+    writeln!(out, "| | CRAP | CC | Cov % | Function | Location |")?;
+    writeln!(out, "|---|---:|---:|---:|---|---|")?;
+    for entry in entries {
+        let grade = Grade::of(entry.crap, threshold);
+        let cov = match entry.coverage {
+            Some(p) => format!("{p:.1}"),
+            None => "—".to_string(),
+        };
+        let func = linkify(
+            format!("`{}`", entry.function),
+            links,
+            &entry.file,
+            entry.line,
+        );
+        let loc = linkify(
+            format!("`{}:{}`", entry.file.display(), entry.line),
+            links,
+            &entry.file,
+            entry.line,
+        );
+        writeln!(
+            out,
+            "| {} | {:.1} | {} | {} | {} | {} |",
+            grade.icon(),
+            entry.crap,
+            entry.cyclomatic as usize,
+            cov,
+            func,
+            loc,
+        )?;
+    }
+    Ok(())
+}
+
+/// Render a GFM markdown table. Coverage bars are replaced by plain
+/// percentages so the table renders correctly in any markdown renderer.
+pub(crate) fn render_markdown(
+    entries: &[CrapEntry],
+    threshold: f64,
+    links: Option<&SourceLinks>,
+    out: &mut dyn Write,
+) -> Result<()> {
+    write_pr_comment_marker(out)?;
+    if entries.is_empty() {
+        writeln!(out, "_No functions found._")?;
+        return Ok(());
+    }
+    let crappy = super::crappy_count(entries, threshold);
+    write_markdown_absolute_heading(crappy, threshold, out)?;
+    write_per_crate_markdown(entries, threshold, out)?;
+    write_markdown_entries_table(entries, threshold, links, out)?;
+    write_markdown_absolute_summary(crappy, entries.len(), threshold, out)
+}
+
+/// Write the "Removed since baseline" section for markdown output.
+fn write_markdown_removed(
+    removed: &[crate::delta::RemovedEntry],
+    out: &mut dyn Write,
+) -> Result<()> {
+    writeln!(out)?;
+    writeln!(out, "**Removed since baseline:**")?;
+    for r in removed {
+        writeln!(out, "- `{}` (was {:.1})", r.function, r.baseline_crap)?;
+    }
+    Ok(())
+}
+
+fn write_markdown_delta_heading(
+    regressions: usize,
+    out: &mut dyn Write,
+) -> Result<()> {
+    if regressions == 0 {
+        writeln!(out, "## ✅ No CRAP regressions")?;
+    } else {
+        writeln!(out, "## ⚠️ {regressions} CRAP regression(s) detected")?;
+    }
+    writeln!(out)?;
+    Ok(())
+}
+
+fn write_delta_entries_table(
+    entries: &[DeltaEntry],
+    threshold: f64,
+    links: Option<&SourceLinks>,
+    out: &mut dyn Write,
+) -> Result<()> {
+    writeln!(out, "| | CRAP | Δ | CC | Cov % | Function | Location |")?;
+    writeln!(out, "|---|---:|---:|---:|---:|---|---|")?;
+    for de in entries {
+        let e = &de.current;
+        let grade = Grade::of(e.crap, threshold);
+        let cov = e.coverage.map_or("—".to_string(), |p| format!("{p:.1}"));
+        let func = linkify(format!("`{}`", e.function), links, &e.file, e.line);
+        let loc = linkify(
+            format!("`{}:{}`", e.file.display(), e.line),
+            links,
+            &e.file,
+            e.line,
+        );
+        writeln!(
+            out,
+            "| {} | {:.1} | {} | {} | {} | {} | {} |",
+            grade.icon(),
+            e.crap,
+            delta_display(de),
+            e.cyclomatic as usize,
+            cov,
+            func,
+            loc,
+        )?;
+    }
+    Ok(())
+}
+
+fn write_markdown_delta_stats(
+    report: &DeltaReport,
+    out: &mut dyn Write,
+) -> Result<()> {
+    let regressed = report
+        .entries
+        .iter()
+        .filter(|e| e.status == DeltaStatus::Regressed)
+        .count();
+    let improved = report
+        .entries
+        .iter()
+        .filter(|e| e.status == DeltaStatus::Improved)
+        .count();
+    let new = report
+        .entries
+        .iter()
+        .filter(|e| e.status == DeltaStatus::New)
+        .count();
+    let unchanged = report
+        .entries
+        .iter()
+        .filter(|e| e.status == DeltaStatus::Unchanged)
+        .count();
+    writeln!(out)?;
+    writeln!(
+        out,
+        "↑ {regressed} regressed · ↓ {improved} improved · ★ {new} new · · {unchanged} unchanged · — {} removed",
+        report.removed.len(),
+    )?;
+    Ok(())
+}
+
+pub(crate) fn render_delta_markdown(
+    report: &DeltaReport,
+    threshold: f64,
+    links: Option<&SourceLinks>,
+    out: &mut dyn Write,
+) -> Result<()> {
+    write_pr_comment_marker(out)?;
+    if report.entries.is_empty() && report.removed.is_empty() {
+        writeln!(out, "_No functions found._")?;
+        return Ok(());
+    }
+    write_markdown_delta_heading(report.regression_count(), out)?;
+    write_delta_entries_table(&report.entries, threshold, links, out)?;
+    if !report.removed.is_empty() {
+        write_markdown_removed(&report.removed, out)?;
+    }
+    write_markdown_delta_stats(report, out)
+}
